requirements.txt
##
# Licensed under the Apache License, Version 2.0 (the "License"); you may
# not use this file except in compliance with the License. You may obtain
# a copy of the License at
#
#         http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
# License for the specific language governing permissions and limitations
# under the License.
##
decorator==4.0.10
Django==1.11.18
django-model-utils==2.6
functools32==3.2.3.post2
jsonfield==1.0.3
jsonschema==2.5.1
pbr==1.10.0
pyaml==15.8.2
python-dateutil==2.6.0
PyYAML==3.12
requests==2.12.4
six==1.10.0
smmap2==2.0.3
url==0.2.0
wheel==0.24.0
mysqlclient==1.4.6
sqlalchemy==1.3.20
